Julie final

+++ AsyncInn/Controllers/HotelRoomController.cs
@@ -14,12 +14,20 @@ namespace AsyncInn.Controllers
     {
         private readonly HotelMgmtDBContext _context;
 
+        /// <summary>
+        /// sets up the database for hotel rooms
+        /// </summary>
+        /// <param name="context">database details</param>
         public HotelRoomController(HotelMgmtDBContext context)
         {
             _context = context;
         }
 
         // GET: HotelRoom
+        /// <summary>
+        /// calls the main Hotel room page
+        /// </summary>
+        /// <returns>the page</returns>
         public async Task<IActionResult> Index()
         {
             var hotelMgmtDBContext = _context.HotelRoomTable.Include(h => h.Hotel).Include(h => h.Room);
@@ -27,9 +35,15 @@ namespace AsyncInn.Controllers
         }
 
         // GET: HotelRoom/Details/5
-        public async Task<IActionResult> Details(int? id)
+        /// <summary>
+        /// Calls the details page
+        /// </summary>
+        /// <param name="HotelID">which hotelroom</param>
+        /// <param name="RoomNumber">which room number</param>
+        /// <returns>the page with the details</returns>
+        public async Task<IActionResult> Details(int? HotelID, int? RoomNumber)
         {
-            if (id == null)
+            if (HotelID == null)
             {
                 return NotFound();
             }
@@ -37,7 +51,7 @@ namespace AsyncInn.Controllers
             var hotelRoom = await _context.HotelRoomTable
                 .Include(h => h.Hotel)
                 .Include(h => h.Room)
-                .FirstOrDefaultAsync(m => m.HotelID == id);
+                .FirstOrDefaultAsync(m => m.HotelID == HotelID);
             if (hotelRoom == null)
             {
                 return NotFound();
@@ -47,10 +61,14 @@ namespace AsyncInn.Controllers
         }
 
         // GET: HotelRoom/Create
+        /// <summary>
+        /// calls the create page
+        /// </summary>
+        /// <returns>the page to create a hotelroom</returns>
         public IActionResult Create()
         {
             ViewData["HotelID"] = new SelectList(_context.HotelTable, "ID", "Name");
-            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "RoomNumber", "RoomNumber");
+            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "ID", "RoomNumber");
             ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name");
             return View();
         }
@@ -58,37 +76,62 @@ namespace AsyncInn.Controllers
         // POST: HotelRoom/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /// <summary>
+        /// requests the DB to add a Hotelroom
+        /// </summary>
+        /// <param name="hotelRoom">the hotel room details</param>
+        /// <returns>the page with the hotel rooms</returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("HotelID,RoomNumber,RoomID,Rate,PetFriendly")] HotelRoom hotelRoom)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(hotelRoom);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                //or should I try a linq and then if it is found in the table then return a notfound()
+                try
+                {
+                    _context.Add(hotelRoom);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                catch (Exception)
+                {
+                    //this does not work
+                    /*Console.WriteLine("That room number was already used or you are out of range");*/
+                    return NotFound();
+                }
             }
             ViewData["HotelID"] = new SelectList(_context.HotelTable, "ID", "Name", hotelRoom.HotelID);
-            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "RoomNumber", "RoomNumber", hotelRoom.RoomNumber);
+            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "ID", "RoomNumber", hotelRoom.RoomNumber);
             ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name", hotelRoom.RoomID);
             return View(hotelRoom);
         }
 
         // GET: HotelRoom/Edit/5
-        public async Task<IActionResult> Edit(HotelRoom key)
+        /// <summary>
+        /// gets the page to edit the hotel room
+        /// </summary>
+        /// <param name="HotelID">which hotel</param>
+        /// <param name="RoomNumber">which hotel room number</param>
+        /// <returns>the page with the details</returns>
+        public async Task<IActionResult> Edit(int? HotelID, int? RoomNumber)
         {
-            if (key == null)
+            if (HotelID == null)
             {
                 return NotFound();
             }
 
-            var hotelRoom = await _context.HotelRoomTable.FindAsync(key);
+            var hotelRoom = await _context.HotelRoomTable
+                .Include(h => h.Hotel)
+                .Include(h => h.Room)
+                .FirstOrDefaultAsync(m => m.HotelID == HotelID);
             if (hotelRoom == null)
             {
                 return NotFound();
             }
+
             ViewData["HotelID"] = new SelectList(_context.HotelTable, "ID", "Name", hotelRoom.HotelID);
-            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "RoomNumber", "RoomNumber", hotelRoom.RoomNumber);
+            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "ID", "RoomNumber", hotelRoom.RoomNumber);
             ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name", hotelRoom.RoomID);
             return View(hotelRoom);
         }
@@ -96,13 +139,24 @@ namespace AsyncInn.Controllers
         // POST: HotelRoom/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /// <summary>
+        /// makes a request to the database to update the details of the hotel room
+        /// </summary>
+        /// <param name="HotelID">which hotel</param>
+        /// <param name="RoomNumber">which hotel room number</param>
+        /// <param name="hotelRoom">the page with all the hotel rooms</param>
+        /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("HotelID,RoomNumber,RoomID,Rate,PetFriendly")] HotelRoom hotelRoom)
+        public async Task<IActionResult> Edit(int? HotelID, int? RoomNumber, [Bind("HotelID,RoomNumber,RoomID,Rate,PetFriendly")] HotelRoom hotelRoom)
         {
-            if (id != hotelRoom.HotelID)
+            if (HotelID != hotelRoom.HotelID)
             {
                 return NotFound();
+            }
+            if(hotelRoom.RoomNumber == RoomNumber)
+            {
+                return RedirectToAction(nameof(Create)); //asks they try recreate
             }
 
             if (ModelState.IsValid)
@@ -126,15 +180,21 @@ namespace AsyncInn.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["HotelID"] = new SelectList(_context.HotelTable, "ID", "Name", hotelRoom.HotelID);
-            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "RoomNumber", "RoomNumber", hotelRoom.RoomNumber);
+            ViewData["RoomNumber"] = new SelectList(_context.RoomTable, "ID", "RoomNumber", hotelRoom.RoomNumber);
             ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name", hotelRoom.RoomID);
             return View(hotelRoom);
         }
 
         // GET: HotelRoom/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        /// <summary>
+        /// calls the delete page
+        /// </summary>
+        /// <param name="HotelID">which hotel</param>
+        /// <param name="RoomNumber">which hotel room number</param>
+        /// <returns>the page</returns>
+        public async Task<IActionResult> Delete(int? HotelID, int? RoomNumber)
         {
-            if (id == null)
+            if (HotelID == null)
             {
                 return NotFound();
             }
@@ -142,7 +202,7 @@ namespace AsyncInn.Controllers
             var hotelRoom = await _context.HotelRoomTable
                 .Include(h => h.Hotel)
                 .Include(h => h.Room)
-                .FirstOrDefaultAsync(m => m.HotelID == id);
+                .FirstOrDefaultAsync(m => m.HotelID == HotelID);
             if (hotelRoom == null)
             {
                 return NotFound();
@@ -152,16 +212,41 @@ namespace AsyncInn.Controllers
         }
 
         // POST: HotelRoom/Delete/5
+        /// <summary>
+        /// makes a request to the DB to remove an hotel room
+        /// </summary>
+        /// <param name="HotelID">which hotel</param>
+        /// <param name="RoomNumber">which room number</param>
+        /// <returns>the page with the remaining objects</returns>
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int HotelID, int RoomNumber)
         {
-            var hotelRoom = await _context.HotelRoomTable.FindAsync(id);
-            _context.HotelRoomTable.Remove(hotelRoom);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            //if (HotelID == null)
+            //{
+            //    return NotFound();
+            //}
+            var hotelRoom = await _context.HotelRoomTable
+                .Include(h => h.Hotel)
+                .Include(h => h.Room)
+                .FirstOrDefaultAsync(m => m.HotelID == HotelID && m.RoomNumber == RoomNumber);
+            if (hotelRoom == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                _context.HotelRoomTable.Remove(hotelRoom);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
         }
 
+        /// <summary>
+        /// determines if the hotel room was in the db or not
+        /// </summary>
+        /// <param name="id">which hotel room</param>
+        /// <returns>true or false</returns>
         private bool HotelRoomExists(int id)
         {
             return _context.HotelRoomTable.Any(e => e.HotelID == id);
