+++ AsyncInn/Controllers/RoomAmenitiesController.cs
@@ -14,12 +14,20 @@ namespace AsyncInn.Controllers
     {
         private readonly HotelMgmtDBContext _context;
 
+        /// <summary>
+        /// sets up the db
+        /// </summary>
+        /// <param name="context">the db</param>
         public RoomAmenitiesController(HotelMgmtDBContext context)
         {
             _context = context;
         }
 
         // GET: RoomAmenities
+        /// <summary>
+        /// calls the main page
+        /// </summary>
+        /// <returns>the page</returns>
         public async Task<IActionResult> Index()
         {
             var hotelMgmtDBContext = _context.RoomAmenitiesTable.Include(r => r.Amenities).Include(r => r.Room);
@@ -27,9 +35,15 @@ namespace AsyncInn.Controllers
         }
 
         // GET: RoomAmenities/Details/5
-        public async Task<IActionResult> Details(int? id)
+        /// <summary>
+        /// calls the details page
+        /// </summary>
+        /// <param name="RoomID">which room</param>
+        /// <param name="AmenitiesID">which amenity</param>
+        /// <returns>the page</returns>
+        public async Task<IActionResult> Details(int? RoomID, int? AmenitiesID)
         {
-            if (id == null)
+            if (RoomID == null )
             {
                 return NotFound();
             }
@@ -37,7 +51,7 @@ namespace AsyncInn.Controllers
             var roomAmenities = await _context.RoomAmenitiesTable
                 .Include(r => r.Amenities)
                 .Include(r => r.Room)
-                .FirstOrDefaultAsync(m => m.AmenitiesID == id);
+                .FirstOrDefaultAsync(m => m.RoomID == RoomID);
             if (roomAmenities == null)
             {
                 return NotFound();
@@ -47,6 +61,10 @@ namespace AsyncInn.Controllers
         }
 
         // GET: RoomAmenities/Create
+        /// <summary>
+        /// calls the create page
+        /// </summary>
+        /// <returns>the page</returns>
         public IActionResult Create()
         {
             ViewData["AmenitiesID"] = new SelectList(_context.AmenitiesTable, "ID", "Name");
@@ -57,6 +75,11 @@ namespace AsyncInn.Controllers
         // POST: RoomAmenities/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /// <summary>
+        /// makes a request to the db to add a room amenity
+        /// </summary>
+        /// <param name="roomAmenities">what is the new room amenity</param>
+        /// <returns>the room amenities</returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("AmenitiesID,RoomID")] RoomAmenities roomAmenities)
@@ -72,65 +95,16 @@ namespace AsyncInn.Controllers
             return View(roomAmenities);
         }
 
-        // GET: RoomAmenities/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var roomAmenities = await _context.RoomAmenitiesTable.FindAsync(id);
-            if (roomAmenities == null)
-            {
-                return NotFound();
-            }
-            ViewData["AmenitiesID"] = new SelectList(_context.AmenitiesTable, "ID", "Name", roomAmenities.AmenitiesID);
-            ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name", roomAmenities.RoomID);
-            return View(roomAmenities);
-        }
-
-        // POST: RoomAmenities/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("AmenitiesID,RoomID")] RoomAmenities roomAmenities)
-        {
-            if (id != roomAmenities.AmenitiesID)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(roomAmenities);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!RoomAmenitiesExists(roomAmenities.AmenitiesID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["AmenitiesID"] = new SelectList(_context.AmenitiesTable, "ID", "Name", roomAmenities.AmenitiesID);
-            ViewData["RoomID"] = new SelectList(_context.RoomTable, "ID", "Name", roomAmenities.RoomID);
-            return View(roomAmenities);
-        }
-
         // GET: RoomAmenities/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        /// <summary>
+        /// calls the delete page
+        /// </summary>
+        /// <param name="RoomID">which room</param>
+        /// <param name="AmenitiesID">which amenity</param>
+        /// <returns>the page</returns>
+        public async Task<IActionResult> Delete(int? RoomID, int? AmenitiesID)
         {
-            if (id == null)
+            if (RoomID == null)
             {
                 return NotFound();
             }
@@ -138,7 +112,7 @@ namespace AsyncInn.Controllers
             var roomAmenities = await _context.RoomAmenitiesTable
                 .Include(r => r.Amenities)
                 .Include(r => r.Room)
-                .FirstOrDefaultAsync(m => m.AmenitiesID == id);
+                .FirstOrDefaultAsync(m => m.RoomID == RoomID);
             if (roomAmenities == null)
             {
                 return NotFound();
@@ -148,16 +122,37 @@ namespace AsyncInn.Controllers
         }
 
         // POST: RoomAmenities/Delete/5
+        /// <summary>
+        /// makes a request to the db to remove it 
+        /// </summary>
+        /// <param name="RoomID">which room</param>
+        /// <param name="AmenitiesID">which amenity</param>
+        /// <returns>the remining room amenities</returns>
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int RoomID, int AmenitiesID)
         {
-            var roomAmenities = await _context.RoomAmenitiesTable.FindAsync(id);
-            _context.RoomAmenitiesTable.Remove(roomAmenities);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var roomAmenities = await _context.RoomAmenitiesTable
+                .Include(r => r.Amenities)
+                .Include(r => r.Room)
+                .FirstOrDefaultAsync(m => m.RoomID == RoomID);
+            if (roomAmenities == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                _context.RoomAmenitiesTable.Remove(roomAmenities);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
         }
 
+        /// <summary>
+        /// checks if the room amenity is already in the DB
+        /// </summary>
+        /// <param name="id">which room amenity</param>
+        /// <returns>true or false</returns>
         private bool RoomAmenitiesExists(int id)
         {
             return _context.RoomAmenitiesTable.Any(e => e.AmenitiesID == id);
